[Refactor] 프로젝트 이름 변경 및 불필요한 주석 제거 'PokerGame' -> 'PokerGameApp'
